Move middleware option constants
