+++ msgraphcore/constants.py
@@ -5,3 +5,6 @@ REQUEST_TIMEOUT = 100
 CONNECTION_TIMEOUT = 30
 BASE_URL = 'https://graph.microsoft.com/v1.0'
 SDK_VERSION = '0.0.1-0'
+
+# MiddlewareOptions
+AUTH_MIDDLEWARE_OPTIONS = 'AUTH_MIDDLEWARE_OPTIONS'
