+++ msgraphcore/middleware/_middleware.py
@@ -5,7 +5,7 @@ from urllib3 import PoolManager
 
 from .options.auth_middleware_options import AuthMiddlewareOptions
 from .options.middleware_control import MiddlewareControl
-from .options.constants import AUTH_MIDDLEWARE_OPTIONS
+from ..constants import AUTH_MIDDLEWARE_OPTIONS
 
 
 class MiddlewarePipeline(HTTPAdapter):
