--- msgraphcore/middleware/options/constants.py
@@ -1,1 +1,0 @@
-AUTH_MIDDLEWARE_OPTIONS = 'AUTH_MIDDLEWARE_OPTIONS'
